fix(profile): remove modals blinking
