+++ packages/frontend/src/components/profile/remove_account/index.ts
@@ -1,0 +1,3 @@
+import RemoveAccount from './RemoveAccount';
+
+export default RemoveAccount;
