+++ packages/frontend/src/components/profile/remove_account/ui/index.ts
@@ -30,4 +30,3 @@ export const StyledRemoveAccount = styled(Container)`
         }
     }
 `;
-
